# Make explosion reflections optional, default off