--- a/addons/frag/stringtable.xml
+++ b/addons/frag/stringtable.xml
@@ -1,6 +1,17 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project name="ACE">
     <Package name="Frag">
+        <Key ID="STR_ACE_Frag_Module_DisplayName">
+            <English>Fragmentation Simulation</English>
+            <Polish>Symulacja fragmentacji</Polish>
+            <Spanish>Simulación de fragmentación</Spanish>
+            <German>Splittersimulation</German>
+            <Czech>Simulace fragmentů</Czech>
+            <Portuguese>Simulação de fragmentação</Portuguese>
+            <French>Simulation de la fragmentation</French>
+            <Hungarian>Repesz-szimuláció</Hungarian>
+            <Russian>Симуляция осколков</Russian>
+        </Key>
         <Key ID="STR_ACE_Frag_EnableFrag">
             <English>Fragmentation Simulation</English>
             <Polish>Symulacja fragmentacji</Polish>