@esteldunedain 2016-02-07
> While we are visiting frag, perhaps it's an idea to add the categories for the settings in that module?

But none of the settings are client settable; do we need categories in that case?