@Glowbal 2016-02-07
It's useful for the settings export option we have.