@esteldunedain 2016-02-18
@Glowbal this should be ready for merging now.

Hopefully I'll PR a rewritten version of the module for the next release.